requirements.txt
Django==2.2.1
djangorestframework==3.9.4
pytz==2019.1
sqlparse==0.3.0
Pillow==6.0.0
psycopg2-binary==2.8.2

django-allauth==0.36.0
django-rest-auth==0.9.3
django-cors-headers==2.2.0